requirements.txt
fastapi
uvicorn
transformers
torch
huggingface_hub
python-dotenv
requests
pydantic
accelerate
google-genai